# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mobile Shop POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Supplier list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /suppliers/{id}/history:
    get:
      operationId: getSupplierHistory
      tags: [suppliers]
      summary: Get supplier purchase history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"

  # SALES
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        balance:
          type: number
        notes:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, name, balance]
    SupplierInput:
      type: object
      properties:
        name:
          type: string
        company:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        balance:
          type: number
        notes:
          type: ["string", "null"]
      required: [name]

    # SALE
...
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        unit_cost:
          type: number
        total:
          type: number
      required: [id, product_id, product_name, quantity, unit_cost, total]
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        supplier_id:
          type: ["integer", "null"]
        supplier_name:
          type: ["string", "null"]
        total_amount:
          type: number
        paid_amount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        created_at:
          type: string
      required: [id, invoice_no, total_amount, paid_amount]
...